extrapolate monacoCodeEditor to its own file
